feat: Enhance help command detection in main.ts
The commit refactors the way we detect the help command in the main.ts file. Instead of checking if help is present in the keys or values of minimistResult, we are now using JSON.stringify to search the whole object. Additionally, weve added a check for help in process.argv[2], allowing case-insensitive detection.